backend/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-colorfield==0.8.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
